+++ components/App.tsx
@@ -1,9 +1,10 @@
 import firebase from 'firebase/app';
 import React from 'react';
-import { DragDropContext, Droppable, DroppableProvided, DropResult } from 'react-beautiful-dnd';
+import { Droppable, DroppableProvided } from 'react-beautiful-dnd';
 import Item from '../data/data_model/item';
 import Set from '../data/data_model/set';
 import View from '../data/data_model/view';
+import DragDropContextProvider from './context/DragDropContextProvider';
 import DraggableSet, { DragDropType } from './draggables/DraggableSet';
 
 interface IProps {
@@ -31,102 +32,17 @@ const App = ({ currentView, setCurrentView, sets, setSets, items, setItems }: IP
     newSet.save();
   };
 
-  const move = {
-    inSameField: ({
-      sourceIndex,
-      destIndex,
-      draggableId,
-      array,
-    }: {
-      sourceIndex: number;
-      destIndex: number;
-      draggableId: string;
-      array: string[];
-    }) => {
-      const ids = [...array];
-      ids.splice(sourceIndex, 1);
-      ids.splice(destIndex, 0, draggableId);
-      return ids;
-    },
-    toOtherField: ({
-      source,
-      destination,
-      draggableId,
-      sourceArray,
-      destArray,
-    }: {
-      source: any;
-      destination: any;
-      draggableId: string;
-      sourceArray: string[];
-      destArray: string[];
-    }) => {
-      console.log('one to anothoer move');
-      const startIds = [...sourceArray];
-      startIds.splice(source.index, 1);
-      const finishIds = [...destArray];
-      finishIds.splice(destination.index, 0, draggableId);
-      return { startIds, finishIds };
-    },
-  };
-
-  const onDragEnd = ({ source, destination, draggableId, type }: DropResult) => {
-    if (!destination) {
-      return;
-    }
-    if (destination.droppableId === source.droppableId && destination.index === source.index) {
-      return;
-    }
-    if (destination.droppableId === source.droppableId) {
-      if (type === DragDropType.ITEM) {
-        const set = sets[destination.droppableId];
-
-        const updatedItemIds = move.inSameField({
-          sourceIndex: source.index,
-          destIndex: destination.index,
-          draggableId,
-          array: set.itemIds,
-        });
-        setSets(prev => ({ ...prev, [set.id]: set.update(updatedItemIds) }));
-      }
-      if (type === DragDropType.SET) {
-        const updatedSetIds = move.inSameField({
-          sourceIndex: source.index,
-          destIndex: destination.index,
-          draggableId,
-          array: currentView.setIds,
-        });
-        const updatedView = currentView.update(updatedSetIds);
-        setCurrentView(updatedView);
-      }
-    } else {
-      if (type === DragDropType.ITEM) {
-        const sourceSet = sets[source.droppableId];
-        const destSet = sets[destination.droppableId];
-        const { startIds, finishIds } = move.toOtherField({
-          source,
-          destination,
-          draggableId,
-          sourceArray: sourceSet.itemIds,
-          destArray: destSet.itemIds,
-        });
-        setSets(prev => ({
-          ...prev,
-          [sourceSet.id]: sourceSet.update(startIds),
-          [destSet.id]: destSet.update(finishIds),
-        }));
-        return;
-      }
-      throw new Error('Invalid drag drop type');
-    }
-  };
-
   if (!currentView) {
     return <h2>Loading...</h2>;
   }
 
   return (
-    <DragDropContext onDragEnd={onDragEnd}>
+    <DragDropContextProvider
+      currentView={currentView}
+      setCurrentView={setCurrentView}
+      sets={sets}
+      setSets={setSets}
+    >
       <div className="container">
         <div className="main-wrapper">
           <button className="add" type="button" onClick={addSet}>
@@ -199,7 +115,7 @@ const App = ({ currentView, setCurrentView, sets, setSets, items, setItems }: IP
           border: 2px solid #c4c4c4;
         }
       `}</style>
-    </DragDropContext>
+    </DragDropContextProvider>
   );
 };
